(hana) minor changes in the DPLCM
git-svn-id: https://svn.urbansim.org/trunk@11136 1513fa5d-83ce-422e-ae47-6967bb0fb594

+++ urbansim_parcel/models/development_project_proposal_sampling_model.py
@@ -126,7 +126,7 @@ class DevelopmentProjectProposalSamplingModel(Model):
 
         # set status of accepted proposals to 'active'
         self.proposal_set.modify_attribute(name="status_id", data=self.proposal_set.id_active,
-                                          index=array(self.accepted_proposals, dtype='int'))
+                                          index=array(self.accepted_proposals, dtype='int32'))
         # delete all tentative (not accepted) proposals from the proposal set
         self.proposal_set.remove_elements(where(
                     self.proposal_set.get_attribute("status_id") == self.proposal_set.id_tentative)[0])
